requirements.txt
altgraph==0.17.4
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
decorator==5.2.1
Flask==3.1.1
fuzzywuzzy==0.18.0
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyMuPDF==1.25.5
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-magic==0.4.27
pytz==2025.2
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
self==2020.12.3
setuptools==79.0.0
six==1.17.0
tesseract==0.1.3
tzdata==2025.2
Werkzeug==3.1.3
